semi-complete
I wanted to at least upload what was completed (simple check and variant check) and show the progress on shifted check. I am having trouble getting it to function properly. My logic was that initially it should be like the variant check, but with an additional outer/inner loop to track for changes in the x/y axis. Not only does the code take 3-5 business days to run for the shifted_check function, but it is giving false True's and false False's. I will reupload as I work to figure this out, but I wanted to submit something at least.